+++ CHANGELOG.md
@@ -38,6 +38,8 @@ All notable changes to this project will be documented in this file.
 - Enless 600-065
 - Enless 600-232
 
+- Nexelec - Origin
+
 ### Changed
 
 - Unifyed occupancy occupancy to send 1 or 0 if its not an elsys device